LayoutTests/storage/indexeddb/three-setversion-calls-expected.txt

Three setVersion calls are queued correctly - exercises an otherwise-untested code path

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

indexedDB.open('three-setversion-calls')
connection.setVersion("version 0")
indexedDB.deleteDatabase("three-setversion-calls")
connection.setVersion("version 1")
connection.setVersion("version 2")
setVersion 0
setVersion 0 complete
setVersion 1
setVersion 1 complete
setVersion 2
setVersion 2 complete
finishedAllSetVersions = true
Successfully deleted database
PASS finishedAllSetVersions is true
PASS successfullyParsed is true

TEST COMPLETE
